requirements.txt
flask==1.1.4
flask-mysqldb==0.2.0
markupsafe==2.0.1